Move most of the representation and data-extracting function from handler into lib. Fix bug that caused crash if one of the VOI was 'unconfigured'
