+++ handlers/data_access.js
@@ -13,6 +13,7 @@ const {
 } = require('../lib/unit_conversion.js')
 const gf = require('../lib/grib_functions')
 const su = require('../lib/station_utils.js')
+const mvu = require('../lib/measured_values_utils.js')
 
 // Instantiate logger
 const processenv = require('processenv')
@@ -264,7 +265,6 @@ function getWeatherMosmix (WEATHER_DATA_BASE_PATH, voisConfigs) {
 }
 
 // http://localhost:5000/weather-stations/10505/measured-values?quantities=t_2m&from=1597140000000&to=1597226400000
-
 function getMeasuredValues (WEATHER_DATA_BASE_PATH, voisConfigs) {
   const REPORT_DATA_BASE_PATH = path.join(WEATHER_DATA_BASE_PATH, 'weather', 'weather_reports')
 
@@ -300,142 +300,28 @@ function getMeasuredValues (WEATHER_DATA_BASE_PATH, voisConfigs) {
       return voiConfigs
     }
 
-    function ensureVoiConfigsCorrectness (voiConfigs) {
-      _.forEach(voiConfigs, function (voiConfig) {
-        if (_.isNil(_.get(voiConfig, ['report', 'key']))) {
-          res.status(500).send('Received request for REPORT for unconfigured VOI')
-          req.log.warn({ res: res }, 'Received request for REPORT for unconfigured VOI')
-        }
-      })
-    }
-
-    function getTimeSeriesData (voiConfigs, timeseriesDataCollection) {
-      const timeSeriesData = []
-      voiConfigs.forEach(function (voiConfig) {
-        timeSeriesData.push(timeseriesDataCollection[voiConfig.report.key])
-      })
-      return timeSeriesData
-    }
-
-    function formatTimeseriesDataArray (timeseriesDataArray) {
-      const timestampsToRemove = []
-      _.forEach(timeseriesDataArray, function (timeseriesData) {
-        _.forEach(timeseriesData, (item) => {
-          if (!_.isNil(item.value)) {
-            return
-          }
-
-          const betterItem = _.find(timeseriesData, (item2) => {
-            return item2.timestamp === item.timestamp && !_.isNil(item2.value)
-          })
-
-          if (!_.isNil(betterItem)) {
-            timestampsToRemove.push(item.timestamp)
-          }
-        })
-      })
-
-      _.forEach(timeseriesDataArray, function (timeseriesData) {
-        _.remove(timeseriesData, (item) => {
-          return _.includes(timestampsToRemove, item.timestamp) && _.isNil(item.value)
-        })
-      })
-      return timeseriesDataArray
-    }
-
-    function renderMeasuredValuesAsJSON (voiConfigs, timeseriesDataArray) {
-      const result = {
-        description: `Quantities ${vois} measured at station ${sid}`,
-        data: []
-      }
-      for (let i = 0; i < voiConfigs.length; i += 1) {
-        const voiConfig = voiConfigs[i]
-        const dataElement = {
-          label: voiConfig.target.key,
-          unit: voiConfig.target.unit,
-          timeseries: timeseriesDataArray[i]
-        }
-        result.data.push(dataElement)
-      }
-      if (_.isNil(result.data[0].timeseries)) {
-        result.data[0].timeseries = []
-      }
-      return result
-    }
-
-    function getCSVLabels (voiConfigs) {
-      let csvLabels = 'timestamp / ms,'
-      _.forEach(voiConfigs, function (voiConfig) {
-        const key = voiConfig.target.key
-        const unit = voiConfig.target.unit
-        csvLabels += `${key} / ${unit},`
-      })
-      csvLabels += '\n'
-      return csvLabels
-    }
-
-    function getFirstElementsFromTimeseries (timeseriesDataArray) {
-      const result = []
-      _.forEach(timeseriesDataArray, function (item) {
-        result.push(_.first(item))
-      })
-      return result
-    }
-
-    function getLastElementsFromTimeseries (timeseriesDataArray) {
-      // returns from each array in timeseriesDataArray first element.
-      const result = []
-      _.forEach(timeseriesDataArray, function (item) {
-        result.push(_.tail(item))
-      })
-      return result
-    }
-
-    function getValuesAsCSVString (timeStampsAndValues) {
-      // Accepts the results from getFirstElementsFromTimeseries
-      // timeStampsAndValues [{timestamp: 1, value: 2}, {...}]
-      // item {timestamp: 1, value:2}
-      const timestampValue = timeStampsAndValues[0].timestamp
-      let result = `${timestampValue},`
-      _.forEach(timeStampsAndValues, function (item) {
-        result += `${item.value},`
-      })
-      return result
-    }
-
-    function getStringsFromStationDataPayload (timeseriesDataArray) {
-      function helper (timeseries, result) {
-        if (_.isEmpty(getFirstElementsFromTimeseries(timeseries)[0])) {
-          return result
-        }
-        const stringWithValues = getFirstElementsFromTimeseries(timeseries)
-        result.push(getValuesAsCSVString(stringWithValues))
-        return helper(getLastElementsFromTimeseries(timeseries), result)
-      }
-      return helper(timeseriesDataArray, [])
-    }
-
-    function renderMeasuredValuesAsCSV (voiConfigs, timeseriesDataArray) {
-      const csvLabels = getCSVLabels(voiConfigs, timeseriesDataArray)
-      const csvStrings = getStringsFromStationDataPayload(timeseriesDataArray)
-      return csvLabels + csvStrings.join('\n')
+    const voiConfigs = getVoiConfigsAsArray(vois)
+    const checkedVois = mvu.ensureVoiConfigsCorrectness(voiConfigs)
+    
+    if (_.includes(checkedVois, false)) {
+      res.status(500).send('received a request for REPORT for unconfigured VOI')
+      req.log.warn({ res: res }, 'received request for REPORT for unconfigured VOI')
+      return
     }
 
     const timeseriesDataCollection = await csv.readTimeseriesDataReport(REPORT_DATA_BASE_PATH, startTimestamp, endTimestamp, sid)
-    const voiConfigs = getVoiConfigsAsArray(vois)
-    ensureVoiConfigsCorrectness(voiConfigs)
     // timeseriesDataArray is an array for timeseries for each voi: [[{}, {}, {}], [{}, {}, {}]]
     // timeseriesDataArray = [[{str:int, str: null}, {...} <-tmstmp]<-voi,[{...}, {...}]<-voi]
-    const timeseriesDataArray = formatTimeseriesDataArray(getTimeSeriesData(voiConfigs, timeseriesDataCollection))
+    const timeseriesDataArray = mvu.formatTimeseriesDataArray(mvu.getTimeSeriesData(voiConfigs, timeseriesDataCollection))
 
     res.format({
       'application/json': function () {
-        const measuredValues = renderMeasuredValuesAsJSON(voiConfigs, timeseriesDataArray)
+        const measuredValues = mvu.renderMeasuredValuesAsJSON(voiConfigs, timeseriesDataArray, vois, sid)
         res.status(200).send(measuredValues)
       },
 
       'text/csv': function () {
-        const measuredValues = renderMeasuredValuesAsCSV(voiConfigs, timeseriesDataArray)
+        const measuredValues = mvu.renderMeasuredValuesAsCSV(voiConfigs, timeseriesDataArray)
         res.status(200).send(measuredValues)
       },
